requirements.txt
pocketflow>=0.0.1
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
openai>=1.0.0
python-multipart>=0.0.6
pyyaml>=6.0.0
python-dotenv>=1.0.0